不要なファイルの削除とかフォルダの整理
