+++ frontend/src/App.js
@@ -4,11 +4,11 @@ import { BrowserRouter as Router, Route } from "react-router-dom";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 
-import Home from "./components/Home";
-import Games from "./components/Games";
-import GameArticle from "./components/GameArticle";
-import Blog from "./components/Blog";
-import BlogArticle from "./components/BlogArticle";
+import Home from "./components/HomeComponents/Home";
+import Games from "./components/GameComponents/Games";
+import GameArticle from "./components/GameComponents/GameArticle";
+import Blog from "./components/BlogComponents/Blog";
+import BlogArticle from "./components/BlogComponents/BlogArticle";
 
 const App = () => {
   return (
@@ -19,13 +19,13 @@ const App = () => {
         <Route exact path="/Games" component={Games} />
         <Route
           exact
-          path="/Games/GameArticle/:id"
+          path="/Games/:id"
           render={(props) => <GameArticle {...props} />}
         />
         <Route exact path="/Blog" component={Blog} />
         <Route
           exact
-          path="/Blog/Article/:id"
+          path="/Blog/:id"
           render={(props) => <BlogArticle {...props} />}
         />
         <Footer></Footer>
