+++ frontend/src/components/GameComponents/GameArticle.js
@@ -1,32 +1,22 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import ReactMarkdown from "react-markdown";
-// import Article from "./Article";
-import CodeBlock from "./CodeBlock.tsx";
 
-import { Grid, Typography } from "@material-ui/core";
+import { Typography, Grid } from "@material-ui/core";
 
 import { makeStyles } from "@material-ui/core/styles";
 import { MuiThemeProvider } from "@material-ui/core/styles";
-import { theme } from "../theme";
+import { theme } from "../../theme";
 
 const useStyles = makeStyles((theme) => ({
   root: {
     margin: "auto",
     maxWidth: 820,
   },
-  date: {
-    marginTop: theme.spacing(4),
-    // marginBottom: theme.spacing(2),
-  },
   title: {
     fontWeight: "Bold",
-    marginTop: theme.spacing(2),
+    marginTop: theme.spacing(4),
     marginBottom: theme.spacing(2),
-  },
-  author: {
-    // marginTop: theme.spacing(2),
-    // marginBottom: theme.spacing(2),
   },
   img: {
     width: "100%",
@@ -48,43 +38,34 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const BlogArticle = (props) => {
+const GameArticle = (props) => {
   const classes = useStyles();
-  const [articles, setArticles] = useState([]);
+  const [games, setGames] = useState([]);
   useEffect(() => {
     async function fetchData() {
       const res = await axios.get(
-        "http://localhost:8000/api/blog/" + props.match.params.id
+        "http://localhost:8000/api/games/" + props.match.params.id
       );
       console.log(res.data);
-      setArticles(res.data);
+      setGames(res.data);
     }
     fetchData();
   }, []);
-  const markdown = articles.text;
+  const markdown = games.text;
+
   return (
     <div>
       <MuiThemeProvider theme={theme}>
         <Grid container direction="column" className={classes.root}>
-          <Typography variant="body1" className={classes.date}>
-            {articles.date}
-          </Typography>
           <Typography variant="h3" className={classes.title}>
-            {articles.title}
+            {games.title}
           </Typography>
-          <Typography variant="body1" className={classes.author}>
-            Author:{articles.author}
-          </Typography>
-          <img src={articles.sumnail} className={classes.img}></img>
-          <ReactMarkdown
-            source={markdown}
-            className={classes.text}
-            renderers={{ code: CodeBlock }}
-          />
+          <img src={games.sumnail} className={classes.img}></img>
+          <ReactMarkdown source={markdown} className={classes.text} />
         </Grid>
       </MuiThemeProvider>
     </div>
   );
 };
 
-export default BlogArticle;
+export default GameArticle;
