+++ frontend/src/components/GameComponents/GameCard.js
@@ -10,7 +10,7 @@ import {
 import { makeStyles } from "@material-ui/core/styles";
 
 import { MuiThemeProvider } from "@material-ui/core/styles";
-import { theme } from "../theme";
+import { theme } from "../../theme";
 
 const useStyles = makeStyles((theme) => ({
   card: {
