+++ frontend/src/components/HomeComponents/Hero.js
@@ -5,7 +5,7 @@ import { makeStyles } from "@material-ui/core/styles";
 
 import { MuiThemeProvider } from "@material-ui/core/styles";
 
-import { theme } from "../theme";
+import { theme } from "../../theme";
 
 const useStyles = makeStyles((theme) => ({
   hero: {
