+++ frontend/src/components/HomeComponents/Members.js
@@ -1,12 +1,12 @@
 import React from "react";
 import Member from "./Member";
 
-import { Grid, Avatar, Typography, Box } from "@material-ui/core";
+import { Grid, Typography, Box } from "@material-ui/core";
 
 import { makeStyles } from "@material-ui/core/styles";
 
 import { MuiThemeProvider } from "@material-ui/core/styles";
-import { theme } from "../theme";
+import { theme } from "../../theme";
 
 const useStyles = makeStyles((theme) => ({
   title: {
